SHO3 - Completed Theme - Shop Added character upgrades

+++ Assets/Scripts/heartScript.cs
@@ -16,7 +16,7 @@ public class heartScript : MonoBehaviour {
 	
 	// Update is called once per frame
      void Update () {
-        HeartUI.sprite = HeartSprites[CharacterScript.lives];
+        HeartUI.sprite = HeartSprites[(int)CharacterScript.lives];
 
     }
 }
